헤더 로그인상태 모듈화
