+++ public/assets/detail.css
@@ -16,10 +16,11 @@ body {
   background: white;
 }
 .content {
+  margin-top: 50px;
   display: flex;
   justify-content: center;
 
-  align-items: center;
+  /* align-items: center; */
   width: 100%;
 }
