+++ public/javascript/detail.js
@@ -1,6 +1,18 @@
-import isLoggedIn from "./isLoggedIn.js";
+import { isLoggedIn, updateLoginStatus } from "./isLoggedIn.js";
 import myPage from "./myPageButton.js";
 myPage();
+updateLoginStatus();
+
+async function updateLoginStatus() {
+  const { loggedIn } = await isLoggedIn(); // 로그인 상태 확인
+  const loginStatusElement = document.getElementById("loginStatus");
+  if (loggedIn) {
+    loginStatusElement.textContent = "로그인 완료";
+  } else {
+    loginStatusElement.textContent = "로그인 안됨";
+  }
+}
+updateLoginStatus();
 
 // 페이지 url에서 postId 추출
 const url = window.location.pathname;
