+++ public/javascript/main.js
@@ -1,17 +1,6 @@
-import isLoggedIn from "./isLoggedIn.js";
+import { isLoggedIn, updateLoginStatus } from "./isLoggedIn.js";
 import myPage from "./myPageButton.js";
 myPage();
-
-async function updateLoginStatus() {
-  const data = await isLoggedIn(); // 로그인 상태 확인
-  const loggedIn = data.isLoggedIn; // 로그인 상태 확인
-  const loginStatusElement = document.getElementById("loginStatus");
-  if (loggedIn) {
-    loginStatusElement.textContent = "로그인 완료";
-  } else {
-    loginStatusElement.textContent = "로그인 안됨";
-  }
-}
 updateLoginStatus();
 
 const searchBooks = () => {
@@ -112,7 +101,6 @@ function generatePostCards(posts, postsList) {
 
 function displayPosts() {
   const postsList = document.getElementById("posts-list");
-  console.log("postsList = ", postsList);
   fetch("api/posts")
     .then(function (response) {
       return response.json();
@@ -195,7 +183,6 @@ displayBestsellers();
 async function initializePage() {
   const result = await isLoggedIn(); // 로그인 상태 확인
   //로그인 상태가 아니면 에러메시지가 온다.
-  console.log(result);
   if (!result.errorMessage) {
     displayFollowingPosts(); // 로그인 상태인 경우 displayFollowingPosts() 실행
   } else {
